Cut out the actual install / uninstall action into its own file
