+++ src/Shimmer.Client/Shimmer.Client.csproj
@@ -120,6 +120,7 @@
   <ItemGroup>
     <Compile Include="BitsManager.cs" />
     <Compile Include="IAppSetup.cs" />
+    <Compile Include="InstallManager.cs" />
     <Compile Include="IUpdateManager.cs" />
     <Compile Include="PostInstallInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
